refactor: split terrain traffic fire and bike-share providers
