+++ scripts/package-boundaries.json
@@ -89,5 +89,47 @@
     "exports": ["./sources/space"],
     "modules": ["src/data/spaceProviderRequests.js"],
     "external": []
+  },
+  "terrain-provider": {
+    "runtime": "node",
+    "exports": ["./server/providers/terrain"],
+    "modules": [
+      "server/providers/terrain.js",
+      "src/data/terrainHeightsProxy.js"
+    ],
+    "external": []
+  },
+  "terrain-source": {
+    "exports": ["./sources/terrain"],
+    "modules": ["src/data/terrainHeightsProxy.js"],
+    "external": []
+  },
+  "traffic-provider": {
+    "runtime": "node",
+    "exports": ["./server/providers/traffic"],
+    "modules": ["server/providers/traffic.js", "src/data/tomtomTiles.js"],
+    "external": []
+  },
+  "traffic-source": {
+    "exports": ["./sources/traffic"],
+    "modules": ["src/data/tomtomTiles.js"],
+    "external": []
+  },
+  "firms-provider": {
+    "runtime": "node",
+    "exports": ["./server/providers/firms"],
+    "modules": ["server/providers/firms.js", "src/data/firmsCsv.js"],
+    "external": []
+  },
+  "gbfs-provider": {
+    "runtime": "node",
+    "exports": ["./server/providers/gbfs"],
+    "modules": ["server/providers/gbfs.js", "src/data/gbfsSource.js"],
+    "external": []
+  },
+  "gbfs-source": {
+    "exports": ["./sources/gbfs"],
+    "modules": ["src/data/gbfsSource.js"],
+    "external": []
   }
 }
